requirements.txt
lactolyse >= 0.9.0a7

django-redis~=4.10.0
django-storages[boto3]~=1.7.1
